moves providers into central place
