+++ src/features/app/index.ts
@@ -1,4 +1,3 @@
 // Export providers
-export { default as AppProviders } from './providers/AppProviders';
 export { default as MigrationScreen } from './screens/MigrationScreen';
