+++ src/features/notifications/index.ts
@@ -1,5 +1,2 @@
-// Export providers
-export { default as NotificationProvider } from './providers/NotificationProvider';
-
 // Export hooks
 export { default as useNotifications } from './hooks/useNotifications';
