--- src/features/purchases/index.ts
@@ -1,2 +1,0 @@
-// Export providers
-export { default as PurchaseProvider } from './providers/PurchaseProvider';
