+++ src/features/themes/hooks/useTheme.ts
@@ -1,0 +1,10 @@
+import { ThemeContext } from "@/app/providers";
+import React from "react";
+
+export const useTheme = () => {
+  const context = React.useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
