+++ src/features/themes/index.ts
@@ -1,4 +1,4 @@
 // Export providers
 export { useAppearance, type Appearance } from './hooks/useAppearance';
-export { ThemeContext, default as ThemeProvider, useTheme } from './providers/ThemeProvider';
+export { useTheme } from './hooks/useTheme';
