+++ structure.txt
@@ -1,0 +1,237 @@
+./src
+├── app
+│   ├── App.tsx
+│   ├── assets
+│   │   └── images
+│   │       ├── cat.png
+│   │       ├── girl-with-dog.png
+│   │       └── vet-examining-dog.png
+│   ├── database
+│   │   ├── context.tsx
+│   │   ├── hoc.ts
+│   │   ├── hooks
+│   │   ├── index.ts
+│   │   ├── migration-utility.ts
+│   │   ├── migrations.ts
+│   │   ├── models
+│   │   │   ├── Assessment.ts
+│   │   │   └── Pet.ts
+│   │   ├── schema
+│   │   └── schema.ts
+│   ├── legacy-realm
+│   │   ├── helper.ts
+│   │   ├── models
+│   │   │   ├── index.ts
+│   │   │   ├── Measurement.ts
+│   │   │   └── Pet.ts
+│   │   └── providers
+│   │       └── MigrationProvider.tsx
+│   ├── localization
+│   │   ├── de
+│   │   │   ├── buttons.json
+│   │   │   ├── calendar.json
+│   │   │   ├── cat.json
+│   │   │   ├── common.json
+│   │   │   ├── dog.json
+│   │   │   ├── index.ts
+│   │   │   ├── measurements.json
+│   │   │   ├── other.json
+│   │   │   ├── settings.json
+│   │   │   ├── tips.json
+│   │   │   └── welcome.json
+│   │   ├── en
+│   │   │   ├── buttons.json
+│   │   │   ├── calendar.json
+│   │   │   ├── cat.json
+│   │   │   ├── common.json
+│   │   │   ├── dog.json
+│   │   │   ├── index.ts
+│   │   │   ├── measurements.json
+│   │   │   ├── other.json
+│   │   │   ├── settings.json
+│   │   │   ├── tips.json
+│   │   │   └── welcome.json
+│   │   ├── es
+│   │   │   ├── buttons.json
+│   │   │   ├── calendar.json
+│   │   │   ├── cat.json
+│   │   │   ├── common.json
+│   │   │   ├── dog.json
+│   │   │   ├── index.ts
+│   │   │   ├── measurements.json
+│   │   │   ├── other.json
+│   │   │   ├── settings.json
+│   │   │   ├── tips.json
+│   │   │   └── welcome.json
+│   │   ├── i18n.ts
+│   │   ├── language.json
+│   │   ├── LanguageDetector.ts
+│   │   └── nl
+│   │       ├── buttons.json
+│   │       ├── calendar.json
+│   │       ├── cat.json
+│   │       ├── common.json
+│   │       ├── dog.json
+│   │       ├── index.ts
+│   │       ├── measurements.json
+│   │       ├── other.json
+│   │       ├── settings.json
+│   │       ├── tips.json
+│   │       └── welcome.json
+│   └── store
+│       └── storageKeys.ts
+├── features
+│   ├── _feature-template
+│   │   └── index.ts
+│   ├── app
+│   │   ├── index.ts
+│   │   ├── providers
+│   │   │   └── AppProviders.tsx
+│   │   └── screens
+│   │       └── MigrationScreen.tsx
+│   ├── assessments
+│   │   ├── components
+│   │   │   ├── AllNotes.tsx
+│   │   │   ├── AssessmentItem.tsx
+│   │   │   ├── AssessmentsCalendar.tsx
+│   │   │   └── NotesModal.tsx
+│   │   ├── helpers
+│   │   │   ├── customTracking.ts
+│   │   │   └── helperFunctions.ts
+│   │   ├── hooks
+│   │   ├── index.ts
+│   │   ├── screens
+│   │   │   ├── AddAssessment.tsx
+│   │   │   ├── AllAssessmentsScreen.tsx
+│   │   │   ├── AllNotesScreen.tsx
+│   │   │   └── EditAssessment.tsx
+│   │   └── types
+│   ├── avatar
+│   │   ├── components
+│   │   │   └── Avatar.tsx
+│   │   └── index.ts
+│   ├── charts
+│   │   ├── components
+│   │   │   ├── AssessmentChart.tsx
+│   │   │   └── WeeklyAssessmentDialog.tsx
+│   │   ├── index.ts
+│   │   ├── types.ts
+│   │   └── utils
+│   │       └── helperFunctions.ts
+│   ├── debug
+│   │   ├── index.ts
+│   │   ├── screens
+│   │   │   └── DebugScreen.tsx
+│   │   └── types
+│   ├── events
+│   │   ├── constants
+│   │   │   └── eventNames.ts
+│   │   ├── index.ts
+│   │   └── types
+│   ├── home
+│   │   ├── components
+│   │   │   └── HomeHeader.tsx
+│   │   ├── hooks
+│   │   ├── index.ts
+│   │   ├── screens
+│   │   │   └── HomeScreen.tsx
+│   │   └── types
+│   ├── navigation
+│   │   ├── AppNavigator.tsx
+│   │   ├── components
+│   │   │   ├── CustomNavigationBar.tsx
+│   │   │   └── SupportDialog.tsx
+│   │   ├── feature-types.ts
+│   │   ├── index.ts
+│   │   ├── navigationOptions.tsx
+│   │   ├── routes.ts
+│   │   └── types.tsx
+│   ├── notifications
+│   │   ├── helpers
+│   │   │   └── helperFunctions.ts
+│   │   ├── hooks
+│   │   │   └── useNotifications.ts
+│   │   ├── index.ts
+│   │   └── providers
+│   │       └── NotificationProvider.tsx
+│   ├── onboarding
+│   │   ├── hooks
+│   │   ├── index.ts
+│   │   ├── screens
+│   │   │   ├── OnboardingScreen.tsx
+│   │   │   └── WelcomeScreen.tsx
+│   │   └── types
+│   ├── pdfExport
+│   │   ├── components
+│   │   │   └── ExportPdf.tsx
+│   │   ├── hooks
+│   │   │   └── useAssessmentExporter.ts
+│   │   └── index.ts
+│   ├── pets
+│   │   ├── components
+│   │   │   └── PetItem.tsx
+│   │   ├── helpers
+│   │   │   └── helperFunctions.ts
+│   │   ├── hooks
+│   │   ├── index.ts
+│   │   ├── screens
+│   │   │   ├── AddPet.tsx
+│   │   │   ├── AssessmentSettings.tsx
+│   │   │   ├── CustomTrackingSettingsScreen.tsx
+│   │   │   ├── EditPet.tsx
+│   │   │   └── NotificationSettings.tsx
+│   │   ├── service
+│   │   ├── types
+│   │   └── types.ts
+│   ├── purchases
+│   │   ├── index.ts
+│   │   └── providers
+│   │       └── PurchaseProvider.tsx
+│   ├── README.md
+│   ├── settings
+│   │   ├── components
+│   │   │   └── Settings.tsx
+│   │   ├── hooks
+│   │   ├── index.ts
+│   │   ├── screens
+│   │   │   └── SettingsScreen.tsx
+│   │   └── types
+│   ├── themes
+│   │   ├── darkTheme.json
+│   │   ├── hooks
+│   │   │   └── useAppearance.ts
+│   │   ├── index.ts
+│   │   ├── lightTheme.json
+│   │   ├── providers
+│   │   │   └── ThemeProvider.tsx
+│   │   └── types.ts
+│   └── tips
+│       ├── components
+│       │   ├── GetStartedTip.tsx
+│       │   ├── TalkToVetTip.tsx
+│       │   └── Tips.tsx
+│       ├── hooks
+│       │   └── useTips.ts
+│       └── index.ts
+├── navigation
+├── providers
+└── support
+    ├── components
+    │   ├── CustomChartDot.tsx
+    │   ├── CustomSwiperPagination.tsx
+    │   ├── PulsatingCircle.tsx
+    │   ├── RatingButtons.tsx
+    │   ├── RatingSlider.tsx
+    │   ├── RatingSliderToolTip.tsx
+    │   └── SegmentedButtons
+    │       ├── SegmentedButtonItem.tsx
+    │       ├── SegmentedButtons.tsx
+    │       └── utils.ts
+    ├── constants.ts
+    └── helpers
+        ├── ColorHelper.ts
+        ├── DateTimeHelpers.ts
+        ├── ImageHelper.ts
+        └── TooltipHelper.ts
+
+83 directories, 152 files
